refactor(hal): rebase onto the substrate's test-tier convention
PR #4 predates the loop substrate (INF-02..06) merged to main since it
was opened: its own tests/hal/ layout and duplicated vcan-availability
check now conflict with the tests/{unit,integration,differential,
property} tier convention and the shared vcan_channel fixture/
requires_vcan marker in tests/conftest.py.

- tests/hal/test_hal_errors.py split: the two config-validation cases
  that never touch a socket move to tests/unit/test_hal_config_errors.py
  (T1, no vcan needed); the two lifecycle/capability cases that do move
  to tests/integration/test_hal_lifecycle_errors.py (T2).
- test_hal_abstraction.py, test_hal_boundaries.py move to
  tests/integration/ with pytestmark = pytest.mark.requires_vcan,
  dropping their own vcan_channel/skip logic in favour of the shared one.
- test_hal_real_socketcan.py moves to tests/integration/ and gains
  pytest.mark.requires_hardware, so it shows up under the project-wide
  TAPWRIGHT_HARDWARE=1 gate alongside every other real-hardware case
  (its TAPWRIGHT_TEST_REAL_CAN_IFACE env var still selects which
  interface, an orthogonal concern).
- SPDX headers added to bus.py/errors.py/frame.py (FW-REQ-050, not yet
  enforced when PR #4 was originally written).

No behavioural change to hal/bus.py's public API or logic — this is
purely bringing #3's already-solid implementation forward onto the
substrate's conventions. Full local gate green: ruff, ruff format,
mypy, all four guardrail scripts, 93 passed / 37 skipped (skips are all
vcan/hardware-absence on this Windows dev machine, verified by reason
string).

Refs #3
